Initial commit
Proof of concept. server.py script hosts a simple server that listens
for json messages with a data key which will be repeated to clients.
index.html is a javascript client that takes data and puts it on a
webpage.
sender.py sends a random SVG plot to the server.